requirements.txt
streamlit
PyPDF2
Pillow 
reportlab